requirements.txt
-i https://pypi.python.org/simple
amqp==2.4.2
aniso8601==3.0.2
autopep8==1.4.3
babel==2.6.0
backports.csv==1.0.7
billiard==3.6.0.0
bleach==2.1.4
boto3==1.9.130
boto==2.49.0
botocore==1.12.130
braintree==3.49.0
cairocffi==1.0.2
cairosvg==2.3.0
celery[redis]==4.3.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.2.1
culqipy==0.2.6
defusedxml==0.5.0 ; python_version >= '3.0'
diff-match-patch==20181111
dj-database-url==0.5.0
dj-email-url==0.2.0
django-admin-interface==0.10.3
django-adminplus==0.5
django-appconf==1.0.3
django-babel==0.6.2
django-bootstrap4==0.0.8
django-cache-url==3.0.0
django-celery-email==2.0.1
django-celery-results==1.0.4
django-colorfield==0.1.15
django-countries==5.3.3
django-debug-toolbar-request-history==0.0.9
django-debug-toolbar==1.11
django-elasticsearch-dsl==0.5.1
django-environ==0.4.5
django-filter==2.1.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-fsm==2.6.0
django-graphql-jwt==0.2.1
django-impersonate==1.4
django-import-export==1.2.0
django-js-asset==1.2.2
django-measurement==3.0.0
django-mptt==0.9.1
django-phonenumber-field==2.3.1
django-prices-openexchangerates==1.0.1
django-prices-vatlayer==1.0.2
django-prices==1.0.4
django-recaptcha==1.5.0
django-redis==4.10.0
django-render-block==0.5
django-rest-auth==0.9.5
django-ses==0.8.10
django-silk==2.0.0
django-storages==1.7.1
django-templated-email==2.3.0
django-versatileimagefield==1.10
django-webpack-loader==0.6.0
django==2.2
djangorestframework-api-key==0.3.1
djangorestframework-camel-case==1.0.3
djangorestframework-jwt==1.11.0
djangorestframework==3.9.2
docutils==0.14
drf-flex-fields==0.3.5
drf-yasg==1.15.0
elasticsearch-dsl==6.0.1
elasticsearch==6.3.1
et-xmlfile==1.0.1
faker==1.0.5
freezegun==0.3.11
future==0.17.1
google-i18n-address==2.3.4
google-measurement-protocol==1.0.0
gprof2dot==2016.10.13
graphene-django-optimizer==0.3.6
graphene-django==2.2.0
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
html5lib==1.0.1
idna==2.8
inflection==0.3.1
ipaddress==1.0.22
itypes==1.1.0
jdcal==1.4
jinja2==2.10.1
jmespath==0.9.4
jsonfield==2.0.2
kombu==4.5.0
markdown==3.1
markupsafe==1.1.1
maxminddb-geolite2==2018.703
maxminddb==1.4.1
measurement==2.0.1
mpmath==1.1.0
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.2
phonenumberslite==8.10.9
pillow==5.4.1
prices==1.0.0
promise==2.2.1
psycopg2-binary==2.8.1
purl==1.5
pycodestyle==2.5.0
pycparser==2.19
pygments==2.3.1
pyjwt==1.7.1
pyphen==0.9.5
python-dateutil==2.8.0 ; python_version >= '2.7'
python3-openid==3.1.0 ; python_version >= '3.0'
pytz==2019.1
pyyaml==5.1
raven==6.9.0
razorpay==1.2.0
redis==3.2.1
requests-oauthlib==1.2.0
requests==2.21.0
rest-social-auth==2.1.0
ruamel.yaml==0.15.91
rx==1.6.1
s3transfer==0.2.0
singledispatch==3.4.0.3
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
stripe==2.24.1
sympy==1.4
tablib==0.13.0
text-unidecode==1.2
tinycss2==1.0.2
typing==3.6.6
uritemplate==3.0.0
urllib3==1.24.1
uwsgi==2.0.18 ; platform_system != 'Windows'
vine==1.3.0
weasyprint==47
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
